langchain/requirements.txt
langchain
openai
python-dotenv
requests
fastapi
uvicorn
httpx
langchain-community
chromadb
sentence-transformers
pypdf
psycopg2-binary